ППМ % completion fix

+++ as-15-8/as-15-8-ppm/src/main/java/controllers/ArrangementController.java
@@ -189,11 +189,11 @@ public class ArrangementController {
         if (checkUnits == null)
             throw new AS_15_8_PPM_Exception("Ошибка расчёта процента выполнения мероприятия. checkUnits null");
 
-        Integer notPlannedNotRunning = resultsDownloader.getNotPlannedNotRunningResults(arrangement.get().getId());
+        Long notPlannedNotRunning = resultsDownloader.getNotPlannedNotRunningResults(arrangement.get().getId());
         if (notPlannedNotRunning == null)
             throw new AS_15_8_PPM_Exception("Ошибка расчёта процента выполнения мероприятия. Число результатов not RUNNING и PLANNED null");
 
-        int percent = (notPlannedNotRunning* 100)/checkUnits.size();
+        int percent = (int) ((notPlannedNotRunning* 100)/checkUnits.size());
         return percent;
     }
